feat: make openssl prime gen non-breaking change
TICKET: WP-1794
